Harden worktree metadata and branch rows
