+++ src/app/components/sidebar/project-work/project-work-block.tsx
@@ -7,7 +7,8 @@ import { useDismissibleLayer } from '../../../hooks/useDismissibleLayer'
 import type { Project, Thread, View } from '../../../types'
 import { appToneSubtleClass, appTypeMetaClass } from '../../../ui/classes'
 import { cn } from '../../../utils/cn'
-import { ProjectCompactBranchGroups, ProjectExpandedBranchGroups } from './branch-thread-groups'
+import { ProjectExpandedBranchGroups } from './branch-thread-groups'
+import { ProjectCompactBranchGroups } from './compact-branch-groups'
 import { NewThreadMenu } from './new-thread-menu'
 import { ProjectWorkActionsMenu } from './project-work-actions-menu'
 import { ProjectRenameField } from './project-work-fields'
